move project package from org.cocos.javascript to com.cocos.game
