+++ templates/template-link/platforms/android/app/AndroidManifest.xml
@@ -18,7 +18,7 @@
                    android:value="cocos" />
 
         <activity
-            android:name="org.cocos2dx.javascript.AppActivity"
+            android:name="com.cocos.game.AppActivity"
             android:screenOrientation="landscape"
             android:configChanges="orientation|keyboardHidden|screenSize|screenLayout"
             android:label="@string/app_name"
