+++ templates/template-link/platforms/android/instantapp/AndroidManifest.xml
@@ -21,7 +21,7 @@
         <meta-data android:name="android.app.lib_name"
                    android:value="cocos" />
         <activity
-                android:name="org.cocos2dx.javascript.AppActivity"
+                android:name="com.cocos.game.InstantActivity"
                 android:screenOrientation="landscape"
                 android:configChanges="orientation|keyboardHidden|screenSize|screenLayout"
                 android:label="@string/app_name"
