+++ templates/template-link/platforms/android/src/com/cocos/game/SDKWrapper.java
@@ -23,15 +23,14 @@
  THE SOFTWARE.
  ****************************************************************************/
 
-package org.cocos2dx.javascript;
+package com.cocos.game;
 
 import android.content.Context;
 import android.content.Intent;
 import android.content.res.AssetManager;
 import android.content.res.Configuration;
-import android.opengl.GLSurfaceView;
 import android.os.Bundle;
-import org.cocos2dx.javascript.service.SDKClass;
+import com.cocos.game.service.SDKClass;
 import org.json.*;
 import java.io.BufferedReader;
 import java.io.IOException;
