+++ templates/template-link/platforms/android/src/com/cocos/game/service/SDKInterface.java
@@ -1,4 +1,4 @@
-package org.cocos2dx.javascript.service;
+package com.cocos.game.service;
 
 import android.content.Context;
 import android.content.Intent;
